Added support for UART and removed the bug

+++ 2.3-1/src/c/hardware/avr/includes/AVRPeripheralADC.h
@@ -17,7 +17,7 @@ extern "C" {
 //Function prototypes
 uint8 u8AVRADCSetups(uint8 prescaler, uint8 adc_ref);
 
-uint16 u8AVRReadADCs(uint8 channel);
+uint8 u8AVRReadADCs(uint8 channel);
 
 
 #ifdef  __cplusplus
